requirements/test.txt
-r prod.txt

django-debug-toolbar==4.4.6
flake8-builtins==2.5.0
flake8-commas==4.0.0
flake8-dunder-all==0.4.1
flake8-eradicate==1.5.0
flake8-expression-complexity==0.0.11
flake8-import-order==0.18.2
flake8-quotes==3.4.0
flake8-return==1.2.0
flake8-use-pathlib==0.3.0
parameterized==0.9.0
pep8-naming==0.14.1